trying small box

+++ bonus/src/player/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: escura <escura@student.42wolfsburg.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/19 11:57:44 by escura            #+#    #+#             */
-/*   Updated: 2024/08/24 18:38:47 by marvin           ###   ########.fr       */
+/*   Updated: 2024/08/24 16:47:40 by escura           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ int view_lane_distance(float x1, float y1, float angle)
     float player_angle = player()->angle;
     
     float raw_distance = distance(x1, y1, x2, y2);
-    int adjusted_distance = raw_distance * cos(player_angle - angle);
+    float adjusted_distance = raw_distance * cos(player_angle - angle);
 
     return adjusted_distance;
 }
